REFACTOR: Merge ssllib.c into ssl4pl.c. Move BIO functions to cryptolib.c.
Merging ssllib.c with ssl4pl.c allows more direct access to important
configuration and OpenSSL structures, shortening the code and making
it easier to understand.

From an architectural perspective, the indirection due to ssllib.c
would have been justified if _either_:

   - the ssllib.h data structures were completely opaque (they weren't)
   - *or* ssllib.c did not rely on specifics of SWI-Prolog (it did).

This commit also removes example programs that are now no longer applicable.
